+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.2.2
+
+- Sanitise `&` and `:` as escape charcters in TeX output.
+- Remove space for the title if no title information is present. Game moves will start at the top of the page.
+
 ## v0.2.1
 
 - Fixed 'undefined' appearing in PDF when custom headers are used, but where some are left blank.